--- src/BoilerPlate.js
@@ -1,14 +1,0 @@
-import React from "react";
-import { View, Text, StyleSheet } from "react-native";
-
-// this is just a boilerplate code so it's easy to copy/paste when a new screen or component is added
-
-export default function BoilerPlate() {
-  return (
-    <View>
-      <Text>Boilerplate code</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({});
